move and rename the metadata form component. #1793
